pyproject.toml
[project]
name = "mypytestrequesthomework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "allure-pytest>=2.15.0",
    "faker>=37.5.3",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "typing>=3.10.0.0",
]

[dependency-groups]
dev = []
